Move annotations and reasons to meta/v1 package
As they are not really part of the runtime, but rather API elements.
