+++ apis/meta/v1/annotations.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package runtime
+package v1
 
 const (
 	// ReconcileAtAnnotation is the annotation used for triggering a
