+++ apis/meta/v1/go.mod
@@ -1,0 +1,3 @@
+module github.com/fluxcd/pkg/apis/meta/v1
+
+go 1.14
